+++ allennlp_models/generation/models/copynet_seq2seq.py
@@ -66,10 +66,6 @@ class CopyNetSeq2Seq(Model):
         as input. This metric must accept two arguments when called, both
         of type `List[List[str]]`. The first is a predicted sequence for each item
         in the batch and the second is a gold sequence for each item in the batch.
-    normalize_weights : `bool`, optional (default = `True`)
-        If `True` and `weight` is used in the forward pass, the weights will be normalized
-        to sum to 1 before being applied. Otherwise the weights will be used as-is, although
-        the final loss will be divided by `batch_size`.
     initializer : `InitializerApplicator`, optional
         An initialization strategy for the model weights.
     """
@@ -87,7 +83,6 @@ class CopyNetSeq2Seq(Model):
         target_namespace: str = "target_tokens",
         tensor_based_metric: Metric = None,
         token_based_metric: Metric = None,
-        normalize_weights: bool = True,
         initializer: InitializerApplicator = InitializerApplicator(),
     ) -> None:
         super().__init__(vocab)
@@ -99,7 +94,6 @@ class CopyNetSeq2Seq(Model):
             self.vocab._padding_token, self._target_namespace
         )
         self._copy_index = self.vocab.add_token_to_namespace(copy_token, self._target_namespace)
-        self._normalize_weights = normalize_weights
 
         self._tensor_based_metric = tensor_based_metric or BLEU(
             exclude_indices={self._pad_index, self._end_index, self._start_index}
@@ -191,10 +185,12 @@ class CopyNetSeq2Seq(Model):
             A tensor of shape `(batch_size, target_sequence_length)` which indicates which
             tokens in the target sequence match tokens in the source sequence.
         weight : `torch.Tensor`, optional (default = `None`)
-           A optional tensor of shape `(batch_size,)` or `(batch_size, 1)` which determines
-           how to weight each instance in the batch when calculating the loss.
-           The default of `None` is equivalent to `weights = torch.tensor([1.0] * batch_size)`,
-           which results in a simple unweighted (or equal-weighted) average.
+            A optional tensor of shape `(batch_size,)` or `(batch_size, 1)` which determines
+            how to weight each instance in the batch when calculating the loss.
+            The default of `None` is equivalent to `weights = torch.tensor([1.0] * batch_size)`.
+            The final loss is calculated as `-(weight * log_likelihood).sum() / batch_size`,
+            where `log_likelihood` is a tensor of shape `(batch_size,)`, representing the overall
+            log likelihood of the gold target sequence.
 
         # Returns
 
@@ -546,8 +542,7 @@ class CopyNetSeq2Seq(Model):
             # shape: (batch_size,)
             if len(weight.shape) > 1:
                 weight = weight.squeeze()
-            denominator = batch_size if not self._normalize_weights else weight.sum()
-            loss = -(weight * log_likelihood).sum() / denominator
+            loss = -(weight * log_likelihood).sum() / batch_size
 
         return {"loss": loss}